+++ Common/src/main/java/fuzs/puzzleslib/api/util/v1/CodecExtras.java
@@ -40,6 +40,18 @@ public final class CodecExtras {
                 DataResult.success(listTag == dynamic.getValue() ? listTag.copy() : listTag) :
                 DataResult.error(() -> "Not a list tag: " + tag);
     }, (ListTag listTag) -> new Dynamic<>(NbtOps.INSTANCE, listTag.copy()));
+    /**
+     * Copied from Minecraft 26.1.
+     */
+    public static final Codec<Integer> RGB_COLOR_CODEC = Codec.withAlternative(Codec.INT,
+            ExtraCodecs.VECTOR3F,
+            v -> ARGB.colorFromFloat(1.0F, v.x(), v.y(), v.z()));
+    /**
+     * Copied from Minecraft 26.1.
+     */
+    public static final Codec<Integer> ARGB_COLOR_CODEC = Codec.withAlternative(Codec.INT,
+            ExtraCodecs.VECTOR4F,
+            v -> ARGB.colorFromFloat(v.w(), v.x(), v.y(), v.z()));
 
     private CodecExtras() {
         // NO-OP
@@ -207,5 +219,18 @@ public final class CodecExtras {
         Function<String, E> function = Arrays.stream(enums)
                 .collect(ImmutableMap.toImmutableMap(keyFunction, Function.identity()))::get;
         return Codec.stringResolver(keyFunction, function);
+    }
+
+    /**
+     * Copied from Minecraft 26.1.
+     */
+    public static <I, E> Codec<E> idResolverCodec(Codec<I> value, Function<I, @Nullable E> fromId, Function<E, @Nullable I> toId) {
+        return value.flatXmap(id -> {
+            E element = fromId.apply((I) id);
+            return element == null ? DataResult.error(() -> "Unknown element id: " + id) : DataResult.success(element);
+        }, e -> {
+            I id = toId.apply((E) e);
+            return id == null ? DataResult.error(() -> "Element with unknown id: " + e) : DataResult.success(id);
+        });
     }
 }